tagger/reviewer details added to tags
